Remove the pinging part of culture-cron, to prepare to run it on toby.
We don't need an entire GCP project just to run this one tiny script.
Let's run it on toby where we run a bunch of other cron jobs.  (We
could also run it as a cloud function, but that's also a lot of
overhead for what we need!)

Issue: none

Test plan:
Fingers crossed, will install this on toby and see...